updated common dto's
