+++ src/module/scraper/booking/booking.dto.ts
@@ -59,68 +59,6 @@ export class BookingRunJobResponseDto implements IPlatformRunJobResponse {
   propertyId?: string;
 }
 
-export class BookingStopJobRequestDto implements IPlatformStopJobRequest {
-  @ApiProperty({
-    example: '507f1f77bcf86cd799439011',
-    description: 'MongoDB ObjectId of the job to stop',
-  })
-  jobId: string;
-}
 
-export class BookingStopJobResponseDto implements IPlatformStopJobResponse {
-  @ApiProperty({ example: 200 })
-  status: number;
 
-  @ApiProperty({ example: 'Booking scraping job stopped successfully' })
-  message: string;
 
-  @ApiProperty({ example: '507f1f77bcf86cd799439011' })
-  jobId: string;
-
-  @ApiProperty({ example: 'Cancelled' })
-  finalStatus: string;
-}
-
-export class BookingRerunFailedJobRequestDto implements IPlatformRerunFailedJobRequest {
-  @ApiProperty({
-    example: '507f1f77bcf86cd799439011',
-    description: 'MongoDB ObjectId of the failed or cancelled job to rerun',
-  })
-  jobId: string;
-
-  @ApiProperty({
-    example: '01/01/2024',
-    description: 'Start date for booking scraping (MM/DD/YYYY format)',
-  })
-  startDate: string;
-
-  @ApiProperty({
-    example: '01/31/2024',
-    description: 'End date for booking scraping (MM/DD/YYYY format)',
-  })
-  endDate: string;
-}
-
-export class BookingRerunFailedJobResponseDto implements IPlatformRerunFailedJobResponse {
-  // TODO - complete rerun logic
-  @ApiProperty({ example: 200 })
-  status: number;
-
-  @ApiProperty({ example: 'Failed or cancelled booking job rerun completed successfully' })
-  message: string;
-
-  @ApiProperty({ example: '507f1f77bcf86cd799439011' })
-  jobId: string;
-
-  @ApiProperty({ example: 'Failed' })
-  originalStatus: string;
-
-  @ApiProperty({ example: 'Completed' })
-  finalStatus: string;
-
-  @ApiProperty({ example: 2 })
-  retryAttempt: number;
-
-  @ApiProperty({ type: JobProgressDto })
-  progress: JobProgressDto;
-}
